test(embed): consolidate guard tests into the existing suites
Addresses review feedback that the change was mostly test code, spread over two
new files of a little over 1,000 lines that also covered behaviour this change
does not touch.

Deletes `test_lancedb_incomplete_index_guard.py` and moves its cases to the
suites that already own those surfaces: the pipeline-writer cases to
`test_lancedb_write_policy.py`, the collection-writer cases to
`test_lancedb_collections.py`, and the multimodal row-loss cases to
`test_multimodal_embed.py`. `test_embed_engine_failure_propagation.py` keeps
only the runtime classifier, which has no existing home.

Drops 11 cases that asserted pre-existing untouched behaviour rather than this
change: numpy and malformed-value handling, the collection path's existing
silent skips, a duplicate wrong-length assertion already covered by the
`on_bad_vectors` case, the fatal-set contents, the cause-chain cycle walk, and
three import-topology pins on modules this change does not modify.

What stays is the pair that matters: for each guard, a case that fails on the
unpatched tree, and a case asserting the guard does not fire - the endpoint
path, wrong-length vectors under the configured policy, blank-text and
canonical image rows, `None` and absent embeddings, all-zero vectors, empty
numpy arrays, image-free chunks, and healthy batches.

Net 721 deletions against 430 insertions. Local run of the five affected files:
195 passed, 3 skipped.
