+++ nemo_retriever/tests/test_embed_engine_failure_propagation.py
@@ -286,52 +286,6 @@ def test_engine_lifecycle_classifier_rejects_unrelated_failures() -> None:
     assert runtime._is_engine_lifecycle_failure(RuntimeError(ENGINE_INIT_FAILED))
 
 
-def test_the_fatal_set_stays_narrow() -> None:
-    """Pins the exact fatal set, so widening it is a deliberate edit.
-
-    Each name here has a cited reason in ``runtime.py``. ``EngineDeadError`` is
-    "Unrecoverable" in vLLM's own definition; ``LocalEmbedderReturnedNothingError``
-    and ``LocalEmbedderRowsLostError`` are raised by this repo, each in one place
-    with one meaning. ``OutOfMemoryError`` and ``EngineGenerateError`` were
-    removed because a recoverable backend raises them.
-
-    Cannot run on unmodified HEAD: the set does not exist there.
-    """
-    assert set(runtime._ENGINE_LIFECYCLE_EXC_NAMES) == {
-        "EngineDeadError",
-        "LocalEmbedderReturnedNothingError",
-        "LocalEmbedderRowsLostError",
-    }
-
-
-def test_engine_lifecycle_classifier_visits_a_cyclic_chain_once_per_link() -> None:
-    """The ``seen`` set, not the depth cap, is what stops a cyclic chain.
-
-    Termination alone does not test ``seen``: the depth-5 cap ends the walk
-    either way, so an assertion that the call returns would still pass with
-    ``seen`` deleted. Visit count does distinguish them. A two-link cycle costs
-    two visits with ``seen`` and five without, because the cap then does the
-    stopping.
-
-    Same status as the test above: it cannot run on unmodified HEAD, where the
-    helper does not exist.
-    """
-    visits: list[str] = []
-
-    class _CountingError(RuntimeError):
-        def __str__(self) -> str:
-            visits.append(self.args[0])
-            return str(self.args[0])
-
-    first = _CountingError("first")
-    second = _CountingError("second")
-    first.__context__ = second
-    second.__context__ = first
-
-    assert not runtime._is_engine_lifecycle_failure(first)
-    assert visits == ["first", "second"]
-
-
 # ---------------------------------------------------------------------------
 # Known gap: two embed entry points swallow ``LocalEmbedderRowsLostError``.
 #
@@ -344,173 +298,9 @@ def test_engine_lifecycle_classifier_visits_a_cyclic_chain_once_per_link() -> No
 # ---------------------------------------------------------------------------
 
 
-def test_the_shipped_embed_actors_use_the_guarded_runtime_function() -> None:
-    """All three shipped embed actors resolve to the guarded implementation.
-
-    This is the reachability claim the two gap tests below depend on. If it ever
-    fails, the gap stops being theoretical and the swallowing paths must be
-    fixed.
-    """
-    from nemo_retriever.operators.embed import cpu_operator, gpu_operator, operators
-
-    for module in (operators, gpu_operator, cpu_operator):
-        assert module.embed_text_main_text_embed is runtime.embed_text_main_text_embed
-
-
-def test_text_embed_operator_route_swallows_the_rows_lost_error(monkeypatch: pytest.MonkeyPatch) -> None:
-    """Pins the gap in ``operators/embed/text_embed.py``.
-
-    Its ``except BaseException`` around ``model.embed(...)`` converts the raise
-    into ``{"embedding": None}``, which the LanceDB writer treats as a
-    legitimate silent drop. Unprotected today, and off the shipped route.
-
-    Passes before and after this change: it asserts pre-existing behaviour.
-    """
-    from nemo_retriever.operators.embed import text_embed
-
-    class _LosingEmbedder:
-        def embed(self, texts: list[str], **_kwargs: Any) -> list:
-            raise LocalEmbedderRowsLostError(lost=1, total=len(texts), embedder="_LosingEmbedder")
-
-    out_df = text_embed.embed_text_1b_v2(_batch(), model=_LosingEmbedder(), inference_batch_size=2)
-
-    payloads = list(out_df["text_embeddings_1b_v2"])
-    assert [p["embedding"] for p in payloads] == [None] * 4
-    assert all(p["error"]["type"] == "LocalEmbedderRowsLostError" for p in payloads)
-    # The shape the writer deliberately does not treat as fatal, because
-    # ``text_embed.py`` also writes ``None`` for a legitimate blank-text row.
-    assert not any(p["embedding"] == [] for p in payloads)
-
-
-def test_modality_pipeline_duplicate_swallows_the_rows_lost_error(monkeypatch: pytest.MonkeyPatch) -> None:
-    """Pins the gap in ``common/modality/pipeline/embedding.py``.
-
-    This module exports a function with the *identical* name as the patched one
-    and is re-exported from its package ``__init__``, so it is the higher-risk
-    of the two: a future import-site change could flip to it and remove all
-    three guard layers with no visible diff at the call site.
-
-    Passes before and after this change: it asserts pre-existing behaviour.
-    """
-    from nemo_retriever.common.modality.pipeline import embedding as duplicate
-
-    assert duplicate.embed_text_main_text_embed is not runtime.embed_text_main_text_embed
-
-    monkeypatch.setattr(
-        duplicate,
-        "_embed_group",
-        _raise(LocalEmbedderRowsLostError(lost=2, total=4, embedder="LlamaNemotronEmbed1BV2Embedder")),
-    )
-
-    out_df = duplicate.embed_text_main_text_embed(_batch(), model=object(), inference_batch_size=2)
-
-    payloads = list(out_df["text_embeddings_1b_v2"])
-    assert [p["embedding"] for p in payloads] == [None] * 4
-    assert all(p["error"]["type"] == "LocalEmbedderRowsLostError" for p in payloads)
-
-
-class _StubVLEmbedder:
-    """VL embedder stub whose per-call return length is scripted by the test."""
-
-    def __init__(self, *, images=None, text_image=None, text=None):
-        self._images = images
-        self._text_image = text_image
-        self._text = text
-
-    def embed_images(self, images_b64, *, batch_size=64):
-        return self._images
-
-    def embed_text_image(self, texts, images_b64, *, batch_size=64):
-        return self._text_image
-
-    def embed(self, texts, *, batch_size=64):
-        return self._text
-
-
-def _image_frame(image_values):
-    return pd.DataFrame({"_image_b64": list(image_values), "text": [""] * len(image_values)})
-
-
 # --- the guard must fire: the engine answered short, rows would be lost ---
-
-
-def test_image_mode_short_answer_is_fatal():
-    """A row submitted with an image and answered with nothing must fail the run.
-
-    Without the guard the shortfall is padded with ``None``, which both writers
-    ignore, so the row is dropped and the run still succeeds.
-    """
-    df = _image_frame(["b64-a", "b64-b", "b64-c"])
-    embedder = _StubVLEmbedder(images=[[0.1, 0.2]])  # 1 vector for 3 images
-
-    with pytest.raises(LocalEmbedderRowsLostError) as excinfo:
-        main_text_embed._multimodal_callable_runner(df, embedder=embedder, batch_size=8, embed_modality="image")
-    assert excinfo.value.lost == 2
-    assert excinfo.value.total == 3
-
-
-def test_image_mode_empty_answer_is_fatal():
-    """Zero vectors for a chunk that did submit images is the whole-batch failure."""
-    df = _image_frame(["b64-a", "b64-b"])
-    embedder = _StubVLEmbedder(images=[])
-
-    with pytest.raises(LocalEmbedderRowsLostError):
-        main_text_embed._multimodal_callable_runner(df, embedder=embedder, batch_size=8, embed_modality="image")
-
-
-def test_text_image_mode_short_multimodal_answer_is_fatal():
-    df = pd.DataFrame({"_image_b64": ["b64-a", "b64-b"], "text": ["alpha", "beta"]})
-    embedder = _StubVLEmbedder(text_image=[[0.1, 0.2]])  # 1 vector for 2 paired rows
-
-    with pytest.raises(LocalEmbedderRowsLostError):
-        main_text_embed._multimodal_callable_runner(df, embedder=embedder, batch_size=8, embed_modality="text_image")
-
-
-def test_text_image_mode_short_text_fallback_answer_is_fatal():
-    """The text-only fallback subset is owed one vector per row as well."""
-    df = pd.DataFrame({"_image_b64": ["", ""], "text": ["alpha", "beta"]})
-    embedder = _StubVLEmbedder(text=[[0.1, 0.2]])  # 1 vector for 2 fallback rows
-
-    with pytest.raises(LocalEmbedderRowsLostError):
-        main_text_embed._multimodal_callable_runner(df, embedder=embedder, batch_size=8, embed_modality="text_image")
 
 
 # --- the guard must stay silent: these are legitimate short answers ---
 
 
-def test_image_mode_rows_without_images_do_not_fire_the_guard():
-    """Rows with no image are owed nothing; they get ``None`` by contract.
-
-    This is the false-failure the naive ``if not vecs_list: raise`` would cause.
-    """
-    df = _image_frame(["b64-a", "", "b64-c"])
-    embedder = _StubVLEmbedder(images=[[0.1, 0.2], [0.3, 0.4]])  # 2 vectors for 2 images
-
-    out = main_text_embed._multimodal_callable_runner(df, embedder=embedder, batch_size=8, embed_modality="image")
-    assert out["embeddings"] == [[0.1, 0.2], None, [0.3, 0.4]]
-
-
-def test_image_mode_chunk_with_no_images_at_all_does_not_fire_the_guard():
-    """An image-free chunk submits nothing, so zero vectors back is correct."""
-    df = _image_frame(["", ""])
-    embedder = _StubVLEmbedder(images=[])
-
-    out = main_text_embed._multimodal_callable_runner(df, embedder=embedder, batch_size=8, embed_modality="image")
-    assert out["embeddings"] == [None, None]
-
-
-def test_text_image_mode_mixed_rows_do_not_fire_the_guard():
-    """Paired, text-only and empty rows each get their own correct treatment."""
-    df = pd.DataFrame({"_image_b64": ["b64-a", "", ""], "text": ["alpha", "beta", "   "]})
-    embedder = _StubVLEmbedder(text_image=[[0.1, 0.2]], text=[[0.3, 0.4]])
-
-    out = main_text_embed._multimodal_callable_runner(df, embedder=embedder, batch_size=8, embed_modality="text_image")
-    assert out["embeddings"] == [[0.1, 0.2], [0.3, 0.4], None]
-
-
-def test_image_mode_healthy_batch_does_not_fire_the_guard():
-    df = _image_frame(["b64-a", "b64-b"])
-    embedder = _StubVLEmbedder(images=[[0.1, 0.2], [0.3, 0.4]])
-
-    out = main_text_embed._multimodal_callable_runner(df, embedder=embedder, batch_size=8, embed_modality="image")
-    assert out["embeddings"] == [[0.1, 0.2], [0.3, 0.4]]
